move views from views folder and rename to componets as they will have control logic for each page
