+++ app/src/modules/components/characterComponent.js
