+++ app/src/modules/components/dungeonComponent.js
@@ -1,5 +1,5 @@
-const state = require('../../state/stateController');
-const combat = require('../../combat/combatcontroller');
+const state = require('../state/stateController');
+const combat = require('../combat/combatcontroller');
 
 const player = state.entity.getPlayerState();
 const enemy = state.entity.getNewSkeleton();
