+++ app/src/modules/components/inventoryComponent.js
