+++ app/src/modules/components/shopComponent.js
